requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.2.1
Flask-Login==0.2.11
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
arrow==0.5.4
psycopg2==2.6
